backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=1.60.0
python-dotenv==1.0.1
pydantic==2.9.2
supabase>=2.0.0
google-generativeai>=0.8.0
google-genai>=1.0.0
websockets>=12.0
python-multipart>=0.0.6

# Audio chunking (fallback for large files)
pydub>=0.25.0

# OCR dependencies
pytesseract>=0.3.10
Pillow>=10.0.0
pdfplumber>=0.10.0

Faker>=20.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
reportlab>=4.0.0
requests>=2.31.0
